Update CV.html
Awesome work 😎 

Looks nice, but you could improve a thing or two.. 
If you were to publish this page, the easiest way you could get Google to score it higher, would be to use an h1 
Here's some stiles to make it look like an h2 
https://www.w3schools.com/tags/tag_hn.asp

Second thing.. I see you used 2 naming conventions. " contactInfo "  and "year_activity" . It's usually considered bad practice to use "_" as a naming convention. "_" is used to show sintax, I hope I can remember to explain this on Sunday. 
Then using  " contactInfo " is quite ok. It's called camelCase, and there's a debate if you should use kebab-case instead. I prefer for CSS to use kebab-case since camelCase is used for something else in javascript.. 
You can read more about this here: https://www.freecodecamp.org/news/css-naming-conventions-that-will-save-you-hours-of-debugging-35cea737d849/
(It's kind of an interesting article)

And lastly on naming convention, "title2" this is not a good name :D..  If I had to choose, I'd have to tags. One "title" and one "h2"  as you did here "about interest"

Ok.. and Lastly on the whole thing.. You have some areas with a lot of text. The <p> is the right way to do it. But think about it, if you can maybe make that in another way.